+++ NEWS.md
@@ -1,6 +1,11 @@
 pnpmisc v0.2.0 (development)
 ============================
 
+Breaking changes
+----------------
+
+* The first argument of `layout_preset()` is now `preset` instead of `name` (so we can now pass `layout_grid()`'s new `name` argument in the `...`).
+
 New features
 ------------
 
@@ -26,13 +31,16 @@ New features
 
 * The following enhancements to `layout_grid()`:
 
-  + Adds `angle` and `name` arguments
+  + Adds `angle`, `direction`, and `name` arguments
   + Adds `angle` and `name` columns to returned data frames.
 
 * The following enhancements to `layout_preset()`:
 
-  + Adds `angle` and `name` columns to returned data frames.
+  + New `...` argument(s) passed to `layout_grid()`.
   + New supported `name` value `button_shy_rules_2x2` to support Button Shy's 4-page booklets.
+  + Adds `angle` and `name` columns to returned data frames.
+
+* `layout_name_fn()` can be used to return a function to generate layout names in `layout_grid()`'s `name` argument.
 
 pnpmisc v0.1.1
 ==============